refactor(http): move extract_mulesoft_sources to MulesoftHttpProxyClient subclass (YET-1229)
PR review feedback: ``HttpProxyClient`` is the generic class shared by
multiple connectors. Putting MuleSoft-specific knowledge —
``mule-artifact.json`` layout, ``configs[]`` semantics, ``properties/``
convention — directly on it leaks domain concerns into shared code.

Move the op + helper + constants out of ``http_proxy_client.py`` into
a new ``apollo/integrations/http/mulesoft_proxy_client.py`` defining
``MulesoftHttpProxyClient(HttpProxyClient)``. The subclass adds
``extract_mulesoft_sources``; inherits the full HTTP surface
(``do_request`` / ``download_bytes`` / ``_open_download_response`` /
SSRF guards / etc.) unchanged.

Inheritance over composition (the InformaticaProxyClient pattern):
``InformaticaProxyClient`` composes because it's conceptually distinct
from "HTTP" (session management + V2/V3 API dispatch). MuleSoft's
agent-side concern is just "HTTP plus one MuleSoft-aware post-step on
a single download" — subclassing keeps code-reuse without delegation
boilerplate and matches the actual conceptual relationship.

Wire-up changes:

* ``apollo/agent/proxy_client_factory.py``: new ``_get_proxy_client_mulesoft``
  factory function (lazy-imports the subclass — same pattern as every
  other entry); ``_CLIENT_FACTORY_MAPPING["mulesoft"]`` rewired from
  ``_get_proxy_client_http`` to ``_get_proxy_client_mulesoft``. Comment
  on the registry entry rewritten to explain the subclass relationship
  + that the inherited HTTP surface keeps the existing YET-1130 REST
  calls working unchanged.
* ``tests/test_http_client.py``: reverted to its pre-YET-1229 state —
  ``base64``/``io``/``json``/``zipfile`` imports + the
  ``TestExtractMulesoftSources`` class removed. Lives in the new
  ``tests/test_mulesoft_proxy_client.py`` instead.
* ``tests/test_proxy_client_factory.py``: the existing
  ``test_http_factory_forwards_platform_for_mulesoft`` test pinned
  the OLD wiring (mulesoft routes through ``_get_proxy_client_http``).
  Renamed + rewritten to pin the new wiring: returns a
  ``MulesoftHttpProxyClient``, ``isinstance(client, HttpProxyClient)``
  still holds (sanity-checks the inheritance), platform still flows
  through, and both the new op + inherited HTTP methods are exposed.
* ``tests/test_mulesoft_proxy_client.py`` (new): the 9 cases from the
  previous commit move here verbatim plus one new
  ``test_inherits_http_proxy_client_surface`` that pins the inheritance
  contract — guards against a future refactor that hides parent
  methods on the subclass.

Patch targets in the moved tests adjusted to point at the new module
location (``apollo.integrations.http.mulesoft_proxy_client.tempfile``,
``apollo.integrations.http.mulesoft_proxy_client._logger``);
``get_storage_client`` patch stays at the
``http_proxy_client`` binding because the storage factory is hoisted
there + the inherited ``_get_storage_client`` looks it up there.

82 tests pass across the three touched test files (10 mulesoft + 65
http_client + 7 factory). Pyright clean on all three production files.
